GCS_MAVLink: split parameter functions into a separate file
